Use AzureStorage consistently instead of mix of AzureTable and AzureBlob
There's no way to use table and blob storage separately, so the project that implements Azure storage based templating should always have been named for AzureStorage, not AzureTable. And then the namespaces within it separated out table and blob storage to no meaningful effect because this was always just a single implementation that happens to use both bits of Azure storage.